I removed the default values in ProductSearchRequestDTO that were not necessary. I removed unnecessary using statements. I adjusted the ErrorMessage for OrderBy and Direction. I separated all the classes used for Product into separate folders. I resolved conflicts.
