+++ Levi9.POS.Domain/Common/IProduct/IProductService.cs
@@ -1,6 +1,5 @@
-﻿using Levi9.POS.Domain.DTOs;
-
-namespace Levi9.POS.Domain.Common
+﻿using Levi9.POS.Domain.DTOs.ProductDTOs;
+namespace Levi9.POS.Domain.Common.IProduct
 {
     public interface IProductService
     {
