--- Levi9.POS.Domain/Common/IProductRepository.cs
@@ -1,11 +1,0 @@
-﻿using Levi9.POS.Domain.DTOs;
-using Levi9.POS.Domain.Models;
-
-namespace Levi9.POS.Domain.Common
-{
-    public interface IProductRepository
-    {
-        Task<Product> GetProductByIdAsync(int id);
-        Task<IEnumerable<Product>> SearchProductsAsync(int page,string name, string orderBy, string direction);
-    }
-}
