+++ Levi9.POS.Domain/DTOs/ProductDTOs/ProductSearchRequestDTO.cs
@@ -1,8 +1,8 @@
-﻿namespace Levi9.POS.Domain.DTOs
+﻿namespace Levi9.POS.Domain.DTOs.ProductDTOs
 {
     public class ProductSearchRequestDTO
     {
-        public int Page { get; set; } = 1;
+        public int Page { get; set; }
         public string Name { get; set; }
         public string OrderBy { get; set; }
         public string Direction { get; set; }
