+++ Levi9.POS/Mapper/MappingProfile.cs
@@ -1,7 +1,6 @@
 ﻿using AutoMapper;
 using Levi9.POS.Domain.DTOs;
 using Levi9.POS.Domain.Models;
-using Levi9.POS.WebApi.Request;
 using Levi9.POS.WebApi.Response;
 
 namespace Levi9.POS.WebApi.Mapper
@@ -9,9 +8,7 @@ namespace Levi9.POS.WebApi.Mapper
     public class MappingProfile : Profile
     {
         public MappingProfile()
-        {
-            CreateMap<Product, ProductDTO>();
-            CreateMap<ProductDTO, ProductResponse>();
+        {           
             CreateMap<Document, DocumentDTO>()
                 .ForMember(dest => dest.DocumentItems, opt => opt.MapFrom(src => src.ProductDocuments))
                 .ReverseMap();
@@ -23,7 +20,6 @@ namespace Levi9.POS.WebApi.Mapper
                 .ForMember(dest => dest.Items, opt => opt.MapFrom(src => src.DocumentItems))
                 .ReverseMap();
             CreateMap<DocumentItemDTO, DocumentItemResponse>().ReverseMap();
-            CreateMap<ProductSearchRequest,ProductSearchRequestDTO>();
         }
     }
 }
