+++ Levi9.POS/Request/ProductRequest/ProductSearchRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Levi9.POS.WebApi.Request
+namespace Levi9.POS.WebApi.Request.ProductRequest
 {
     public class ProductSearchRequest
     {
